[MAJOR-9]
* moved the Stats and VisualWrapper from PyStats.py to _PyStats.py file so that the main file is cleaner
* some minor changes in _PyStats.py
